% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_est.R
\name{generate_k}
\alias{generate_k}
\title{Generate initial values for states for uncertain cells}
\usage{
generate_k(u.obs, s.obs, empirical, n_inits = 50, type = "min", plot = TRUE)
}
\arguments{
\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{empirical}{output from \code{get_empirical}.}

\item{n_inits}{number of initializations (chains) to generate for.}

\item{type}{a character string indicating which method is used to generate states. One of 'centre' or 'min'.
If 'centre', the probability to generate states is based on the distance to the centre points, otherwise it is based on the points closest to the steady states.}

\item{plot}{if TRUE, plot the centre points or closest points.}
}
\value{
a matrix of initial values. Each column corresponds to one set of initial values, with rows corresponding to cells.
}
\description{
The function generates initial values for states for uncertains that will be used when running multiple chains, based on two different ways.
}
\examples{
k.inits1 <- generate_k(u.obs = u.obs, s.obs = s.obs, empirical = empirical, n_inits = 50, type = 'centre')
}
