% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_est.R
\name{get_empirical}
\alias{get_empirical}
\title{Derive empirical estimates for parameters}
\usage{
get_empirical(u.obs, s.obs, u.quant, s.quant, alpha='max', epsilon=1e-3, plot=TRUE)
}
\arguments{
\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{u.quant}{a vector of two values for lower and upper quantiles to find empirical repression and induction steady states for unspliced counts.}

\item{s.quant}{a vector of two values for lower and upper quantiles to find empirical repression and induction steady states for spliced counts.}

\item{alpha}{a character string indicating which method is used to calculate empirical alpha. One of 'max' or 'avg'.
If 'max', empirical value is based on the maximum \eqn{u} or \eqn{s}, otherwise it is based on the average.}

\item{epsilon}{a small positive value used to compute empirical \eqn{\tau}.}

\item{plot}{if TRUE, plot empirical \eqn{\gamma} and empirical states.}
}
\value{
The output contains the following items:
\item{tau.hat}{a vector of empirical \eqn{\tau} for all cells.}
\item{k.hat}{a vector of empirical states \eqn{k} for all cells, consisting of two states (1 for induction, 2 for repression).}
\item{k.hat.4}{a vector of empirical states \eqn{k} for all cells, consisting of four states (3 for repression steady state, 4 for induction steady state).}
\item{params}{a matrix of empirical estimates for rates, switching point, initial condition (including its prior hyper-parameters) and variance parameters.}
\item{alpha.type}{a character string indicating which method is used to calculate empirical alpha. One of 'max' or 'avg'.}
}
\description{
The funciton calculates empirical estimates for all parameters (excluding \eqn{\lambda}) based on the model with a constant transcription rate.
}
\examples{
empirical <- get_empirical(u.obs = u.obs, s.obs = s.obs,
  u.quant = c(0.35,0.8), s.quant = c(0.35,0.8), alpha='max')
}
