% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{get_shared_time}
\alias{get_shared_time}
\title{Compute gene-shared latent for each cell}
\usage{
get_shared_time(post_result, n, G)
}
\arguments{
\item{post_result}{output from \code{mcmc_shared_time}.}

\item{n}{number of cells.}

\item{G}{number of genes.}
}
\value{
a list of results for all chains. A \code{mcmc.list} object. Within each list, it contains a matrix of samples for shared latent time.
Each column corresponds to the posterior samples for a single cell.
}
\description{
This function calculates gene-shared latent time for each cell, based on posterior samples from the post-processing step.
}
\examples{
shared_time_samples <- get_shared_time(post_result = post_result, n = 400, G = 26)
}
