% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{mcmc_shared_time}
\alias{mcmc_shared_time}
\title{Implement a post-processing step to infer a gene-shared latent time}
\usage{
mcmc_shared_time(
  x,
  k.mode,
  var.logt,
  p.hat,
  n_chains,
  niter,
  burn_in = 0,
  thinning = 1
)
}
\arguments{
\item{x}{a cell-by-gene matrix. Each entry is the posterior mean of log gene-wise latent time, conditional on the state with a larger posterior probability \code{(k.mode)}.}

\item{k.mode}{a cell-by-gene matrix. Each entry is the state (induction: 1, repression: 2) with a larger posterior probability.}

\item{var.logt}{an array \code{[cell, gene, state]} of posterior variance of log gene-wise latent time, for each state.}

\item{p.hat}{an array \code{[cell, gene, state]} of posterior probability of belonging to each state for each cell in each gene.}

\item{n_chains}{number of chains to run.}

\item{niter}{integer. Total number of MCMC iterations.}

\item{burn_in}{the length of burn-in period during MCMC.}

\item{thinning}{the thinning applied after burn-in period.}
}
\value{
The function returns a list of results for all chains. A \code{mcmc.list} object. Within each list, it contains a matrix of samples for each chain.
The columns correspond to samples for the following parameters: \eqn{\mu_{c,j}, \  \tilde{\sigma}^2_{c,j}, \   \rho_{c,j}, \  -\log(\beta_g)},
for \eqn{j=1,2}, for each cell \eqn{c} and gene \eqn{g}.
}
\description{
This function uses posterior summaries from fitting individual genes to estimate a shared latent time and splicing rate.
}
\examples{
post_result <- mcmc_shared_time(x = y, k.mode = k.mode, var.logt = var.logt, p.hat = p.hat, n_chains = 3,
    niter = 200, burn_in = 100, thinning = 1)
}
