% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{order_probability}
\alias{order_probability}
\title{Calculate posterior probability of one cell before another}
\usage{
order_probability(shared_time_samples)
}
\arguments{
\item{shared_time_samples}{output from \code{get_shared_time}.}
}
\value{
a matrix of probabilities, of dimension \eqn{n \times n} where \eqn{n} is the number of cells. The cells
are ordered by posterior median of shared time from top to bottom, and left to right.
}
\description{
This function computes posterior probability of one cell before another, based on MCMC samples for gene-shared latent time.
}
\examples{
p_order <- order_probability(shared_time_samples = shared_time_samples)
}
