% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_consens.R
\name{plot_distribution_2d}
\alias{plot_distribution_2d}
\title{Plot joint posterior distributions for two parameters}
\usage{
plot_distribution_2d(combined_result, param1, param2)
}
\arguments{
\item{combined_result}{output from \code{result_combine}.}

\item{param1}{name of the first parameter. Must be a character string used in \code{combined_result}.}

\item{param2}{name of the second parameter. Must be a character string used in \code{combined_result}.}
}
\value{
a 2D kernel density estimation for joint posterior distributions.
}
\description{
This function visualizes the joint posterior distributions for a pair of parameters to check correlations.
}
\examples{
plot_distribution_2d(combined_result = combined, param1 = 't0', param2 = 'gamma')
}
