% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_consens.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot fitted phase portrait}
\usage{
plot_fit(combined_result, u.obs, s.obs, thinning = 1, title = NULL, cex = 1)
}
\arguments{
\item{combined_result}{output from \code{result_combine}.}

\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{thinning}{an integer. Only plot for thinned samples.}

\item{title}{a character string for plot title.}

\item{cex}{see \code{par}.}
}
\value{
a phase portrait with samples for fitted lines in grey.
}
\description{
This function plots the posterior samples for fitted phase portrait.
}
\examples{
plot_fit(combined_result = combined, u.obs = u.obs, s.obs = s.obs,
  thinning = 10, title = 'Example', cex = 1)
}
