% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo_consens.R
\name{plot_lp}
\alias{plot_lp}
\title{Plot absolute difference in moving average of log-posterior density}
\usage{
plot_lp(Ws, Ds, mcmc_lp)
}
\arguments{
\item{Ws}{either an integer or a vector of integers. Absolute differences will be computed for each value in \code{Ws}.}

\item{Ds}{a vector of candidate values for depths (number of iterations).}

\item{mcmc_lp}{a matrix of log-posterior density, where the rows correspond to chains and columns correspond to iterations.}
}
\value{
If \code{Ws} is an integer, a line plot showing absolute difference in moving average of log-posterior density between different depths.
If \code{Ws} is a vector, for each value in \code{Ws} a line will be plotted.
}
\description{
This function plot absolute difference in moving average of log-posterior density to select a suitable value for \eqn{D}
(number of iterations) in consensus velocity.
}
\examples{
Ds <- c(1, seq(100,1000,by=100))
plot_lp(Ws = 100, Ds = Ds, mcmc_lp = mcmc_lp_result)
}
