% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_consens.R
\name{plot_predicted_velocity}
\alias{plot_predicted_velocity}
\title{Plot predicted velocity in the phase portrait}
\usage{
plot_predicted_velocity(
  combined_result,
  u.obs,
  s.obs,
  delta = 1,
  obs_ind,
  thinning = 1,
  cex = 1,
  transparency = 1
)
}
\arguments{
\item{combined_result}{output from \code{result_combine}.}

\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{delta}{a time increment to compute cell future position (default to 1).}

\item{obs_ind}{a vector of cell indices for which predicted velocities will be computed.}

\item{thinning}{compute predicted velocity for thinned samples. The posterior mean is also based on thinned samples.}

\item{cex}{see \code{par}.}

\item{transparency}{level of transparency added to the observed values (between 0 and 1).}
}
\value{
a phase portrait overlaid by posterior samples for predicted velocities for selected cells (grey), and posterior mean is shown in red arrows.
}
\description{
This function computes and plots samples for predicted velocities, including posterior mean.
}
\examples{
plot_predicted_velocity(combined_result = combined, u.obs = u.obs, s.obs = s.obs, delta = 1,
    obs_ind = c(1,10,15,20,25,30,35,40,80,100,105,115,120),
    thinning = 10, cex = 0.2, transparency = 0.5)
}
