% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc_multiple}
\alias{ppc_multiple}
\title{Posterior predictive checks based on multiple replicates}
\usage{
ppc_multiple(
  u.obs,
  s.obs,
  combined_result,
  n_replicate,
  prob = c(0.005, 0.995),
  u.quant = c(0.35, 0.8),
  s.quant = c(0.35, 0.8),
  gamma.hat.obs
)
}
\arguments{
\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{combined_result}{output from \code{result_combine}.}

\item{n_replicate}{number of replicates.}

\item{prob}{the target probability content of the credible interval.}

\item{u.quant, s.quant}{a vector of two values for lower and upper quantiles to find
empirical repression and induction steady states for unspliced and spliced counts. Should be the same as used in \code{get_empirical}.}

\item{gamma.hat.obs}{empirical \eqn{\gamma} from the observed data.}
}
\value{
If there are cells whose observed values are not covered by the credible intervals, the first plot will show three panels:
1) a phase portrait where these cells are colored in red, 2) and 3) the observed values and credible intervals will be shown for these cells,
for \eqn{u} and \eqn{s}, separately. If all cells are covered, the first plot will not be shown.

The second plot with three panels will always be shown: 1) comparing empirical \eqn{\gamma} which is represented as the slope of the lines,
2) and 3) comparing quantiles from the replicated datasets against those from the observed data, for \eqn{u} and \eqn{s}, separately. In each panel,
the replicates are shown as grey and the observed data is shown as red.
}
\description{
The function generates multiple replicated datasets, and show observed data that is not covered by
the credible intervals of the replicates. It then compares empirical \eqn{\gamma} and quantiles between
replicates and observed data
}
\examples{
set.seed(4)
ppc_multiple(u.obs = u.obs, s.obs = s.obs, combined_result = combined,
  n_replicate = 1000, prob = c(0.005,0.995), u.quant = c(0.35,0.8),
  s.quant = c(0.35,0.8), gamma.hat.obs = empirical$params$gamma.hat)
}
