% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc_single}
\alias{ppc_single}
\title{Posterior predictive checks based on one single replicate}
\usage{
ppc_single(u.obs, s.obs, combined_result)
}
\arguments{
\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{combined_result}{output from \code{result_combine}.}
}
\value{
two plots comparing generated observations with observed data, for unspliced and spliced counts, respectively.
}
\description{
This function generates one replicated dataset and compares that with the observed data, for a single gene.
}
\examples{
set.seed(4)
ppc_single(u.obs = u.obs, s.obs = s.obs, combined_result = combined)
}
