% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppp_mixed}
\alias{ppp_mixed}
\title{Compute posterior predictive p-values based on mixed predictive distribution}
\usage{
ppp_mixed(u.obs, s.obs, combined_result, n_replicate, prob = 0.9)
}
\arguments{
\item{u.obs}{a vector of observed unspliced counts.}

\item{s.obs}{a vector of observed spliced counts.}

\item{combined_result}{output from \code{result_combine}.}

\item{n_replicate}{number of replicates.}

\item{prob}{a numeric value between 0 and 1. Cells with posterior probabilities of being in a state greater than \code{prob}
will be used to compare their observed counts with generated counts.}
}
\value{
two histograms showing p-values for each state separately.
Also outputs the p-values for \eqn{u} and \eqn{s} for each state, separately.
}
\description{
This function computes posterior predictive p-values based on multiple replicates, by comparing
generated counts with observed counts for \eqn{s} and \eqn{u}, separately. The comparison is done
for each state separately, and based only on cells with a large posterior probability of belonging to that state.
}
\examples{
set.seed(134)
ppp_values <- ppp_mixed(u.obs = u.obs, s.obs = s.obs,
  combined_result = combined, n_replicate = 1000, prob = 0.9)
}
