% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_consens.R
\name{result_combine}
\alias{result_combine}
\title{Combine MCMC samples from multiple chains}
\usage{
result_combine(consensus_result, Width, ind, hyper = FALSE)
}
\arguments{
\item{consensus_result}{output from \code{velo_consensus}.}

\item{Width}{selected value for \eqn{W}. Only the first \code{Width} chains will be saved in the output.}

\item{ind}{sample index at which MCMC samples are saved in the output.}

\item{hyper}{if TRUE, samples for hyper-parameters \eqn{\mu_{\tau,j}} and \eqn{\sigma^2_{\tau,j}} are saved. These will be useful for posterior predictive checks.}
}
\value{
The output returns a list of results for first \code{Width} chains at selected sample index.
Within each list, it contains the following components for each chain:
\item{params}{a matrix of samples for parameters, including rates, \eqn{\lambda}, switching point,
initial condition, variance parameters and hyper-parameter \eqn{p}. Each column corresponds to samples for a variable.}
\item{tau}{a matrix of samples for local time \eqn{\tau}. Each column corresponds to the local time for a single cell.}
\item{mus}{a matrix of samples for \eqn{u(t),s(t)} from the ODE. Each column corresponds to a single cell.}
\item{v, t, k}{a matrix of samples for velocity \eqn{v} from the differential equation for \eqn{s(t)}, for latent time \eqn{t} and for states, respectively.}
\item{hyper}{if \code{hyper}=TRUE, a matrix of samples for hyper-parameters \eqn{\mu_{\tau,j}} and \eqn{\sigma^2_{\tau,j}}, otherwise it is NULL.}
}
\description{
This function combines MCMC samples for parameters from multiple chains into a clearer structure, based on selected \eqn{W} and \eqn{D}.
}
\examples{
combined <- result_combine(consensus_result = consensus_result, Width = 3, ind = 10:100)
}
