% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated unspliced and spliced counts for a single gene}
\format{
A list containing the observed counts, mean counts from the ODE, and parameters to generate the data.
}
\usage{
sim.data
}
\description{
Unspliced and spliced counts for 200 observations (cells) are simulated for a single gene, under the model.
}
\keyword{datasets}
