% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{velocity_pca}
\alias{velocity_pca}
\title{Visualize predicted velocity on PCA space}
\usage{
velocity_pca(
  combined_all_genes,
  post_result,
  u.obs,
  s.obs,
  groups = NULL,
  delta = 1,
  obs_ind,
  thinning = 1,
  cex = 1,
  transparency = 1
)
}
\arguments{
\item{combined_all_genes}{a list of results for all genes. Each item is the output from \code{result_combine} for a single gene.}

\item{post_result}{output from \code{mcmc_shared_time}.}

\item{u.obs}{a cell-by-gene matrix of unspliced counts.}

\item{s.obs}{a cell-by-gene matrix of spliced counts.}

\item{groups}{a vector of strings for colors to show different groups of cells, if available.}

\item{delta}{a time increment to compute cell future position (default to 1).}

\item{obs_ind}{a vector of cell indices for which predicted velocities will be computed.}

\item{thinning}{compute predicted velocity for thinned samples. The posterior mean is also based on thinned samples.}

\item{cex}{see \code{par}.}

\item{transparency}{level of transparancy added to the observed values (between 0 and 1).}
}
\value{
a plot with PC2 against PC1, overlaid by posterior samples for predicted velocities for selected cells (grey),
and posterior mean is shown in red arrows. If \code{groups} is provided, cells are colored based on the groups.
}
\description{
This function computes the posterior mean for the inverse of the splicing rate, which is then used to adjust all
the parameters across genes. After rescaling,  the predicted velocities across genes are computed and then
visualized on the PCA space.
}
\examples{
velocity_pca(combined_all_genes = combined_all_genes, post_result = post_result,
    groups = rep(c('blue','yellow'),each=200),
    u.obs = u.obs, s.obs = s.obs, obs_ind = seq(40,400,by=40), thinning = 1, cex = 0.2)
}
