% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_consens.R
\name{plot_distribution_2d}
\alias{plot_distribution_2d}
\title{Plot joint posterior distributions for two parameters}
\usage{
plot_distribution_2d(combined_result, param1, param2)
}
\arguments{
\item{combined_result}{output from \code{result_combine}.}

\item{param1, param2}{names of the parameters. Must be one of the following:
1) gene-specific parameters 'alpha1_1','gamma','lambda','t0','u01','sigma_u_2','sigma_s_2','p'.
2) latent time 't[1]', 't[2]',...
3) local time 'tau[1]','tau[2]', ...
4) velocity 'v[1]', 'v[2]', ...}
}
\value{
a 2D kernel density estimation for joint posterior distributions.
}
\description{
This function visualizes the joint posterior distributions for a pair of parameters to check correlations.
}
\examples{
plot_distribution_2d(combined_result = combined, param1 = 't0', param2 = 'gamma')
}
