% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo_consens.R
\name{plot_entropy}
\alias{plot_entropy}
\title{Plot absolute difference in entropy}
\usage{
plot_entropy(Ws, Ds, mcmc_k)
}
\arguments{
\item{Ws}{a vector of candidate values for widths (number of chains).}

\item{Ds}{either an integer or a vector of integers. Absolute differences will be computed for each value in \code{Ds}.}

\item{mcmc_k}{a list, each corresponding to the samples of k from a single chain.}
}
\value{
If \code{Ds} is an integer, a line plot showing absolute difference in entropy between different widths.
If \code{Ds} is a vector, for each value in \code{Ds} a line will be plotted.
}
\description{
This function plot absolute difference in entropy to select a suitable value for \eqn{W} (number of chains) in consensus velocity.
}
\examples{
Ws <- c(1, 2, 3)
plot_entropy(Ws = Ws, Ds = 500, mcmc_k = mcmc_k_result)
}
